Changes To MyLib.Convert.ReadLine Method now allows up to one more arguement: a bool value representing whether the requested input is nullable.
Additionally, I reduced the amount of exceptions the method throws by having a delegate declared that handles error correction. Now when the input is null when its not nullable, the delegate is called directly instead of throwing an exception, allowing the program to call the delegate in the catch.

Additionally, I made it use my erasing methods directly instead of using the older outdated versions.